WIP - fix base defence spam
