+++ src/bot/logic/building/antiGroundStaticDefence.ts
@@ -2,6 +2,7 @@ import { GameApi, PlayerData, TechnoRules, Vector2 } from "@chronodivide/game-ap
 import { getPointTowardsOtherPoint } from "../map/map.js";
 import { GlobalThreat } from "../threat/threat.js";
 import { AiBuildingRules, getDefaultPlacementLocation, numBuildingsOwnedOfType } from "./buildingRules.js";
+import { getStaticDefencePlacement } from "./common.js";
 
 export class AntiGroundStaticDefence implements AiBuildingRules {
     constructor(
@@ -15,23 +16,7 @@ export class AntiGroundStaticDefence implements AiBuildingRules {
         playerData: PlayerData,
         technoRules: TechnoRules,
     ): { rx: number; ry: number } | undefined {
-        // Prefer front towards enemy.
-        let startLocation = playerData.startLocation;
-        let players = game.getPlayers();
-        let enemyFacingLocationCandidates: Vector2[] = [];
-        for (let i = 0; i < players.length; ++i) {
-            let playerName = players[i];
-            if (playerName == playerData.name) {
-                continue;
-            }
-            let enemyPlayer = game.getPlayerData(playerName);
-            enemyFacingLocationCandidates.push(
-                getPointTowardsOtherPoint(game, startLocation, enemyPlayer.startLocation, 4, 16, 1.5),
-            );
-        }
-        let selectedLocation =
-            enemyFacingLocationCandidates[Math.floor(game.generateRandom() * enemyFacingLocationCandidates.length)];
-        return getDefaultPlacementLocation(game, playerData, selectedLocation, technoRules, false, 0);
+        return getStaticDefencePlacement(game, playerData, technoRules);
     }
 
     getPriority(
